+++ src/cosmos/cosmosUtils.ts
@@ -88,12 +88,12 @@ export const createCosmosClients = async (
   // eslint-disable-next-line @typescript-eslint/dot-notation
   const queryClient = stargateClient['forceGetQueryClient']()
   // eslint-disable-next-line @typescript-eslint/dot-notation
-  const tendermintClient = stargateClient['forceGetTmClient']()
+  const cometClient = stargateClient['forceGetCometClient']()
 
   return {
     queryClient,
     stargateClient,
-    tendermintClient
+    cometClient
   }
 }